+++ tests/cli/generate/routes/swagger/ErrorsCept.php
@@ -1,4 +1,4 @@
 <?php 
 $I = new CliTester($scenario);
-$I->runShellCommand('bin/rest-deamon generate:routes:swagger no-existed-dir', false);
+$I->runShellCommand('bin/rest-daemon generate:routes:swagger no-existed-dir', false);
 $I->seeResultCodeIs(1);