+++ tests/unit/FreeElephants/AbstractTestCase.php
@@ -2,7 +2,6 @@
 
 namespace FreeElephants;
 
-
 use PHPUnit\Framework\TestCase;
 
 abstract class AbstractTestCase extends TestCase