+++ example/routes.php
@@ -1,11 +1,11 @@
 <?php
 
 use FreeElephants\RestDaemon\Endpoint\Handler\CallableEndpointMethodHandlerWrapper;
-use RestDeamon\Example\Endpoint\Greeting\GetAttributeHandler;
-use RestDeamon\Example\Endpoint\Greeting\GetHandler as GreetingGetHandler;
-use RestDeamon\Example\Endpoint\Greeting\PostHandler;
-use RestDeamon\Example\Endpoint\Index\GetHandler;
-use RestDeamon\Example\Endpoint\Reusable\HelloHandler;
+use RestDaemon\Example\Endpoint\Greeting\GetAttributeHandler;
+use RestDaemon\Example\Endpoint\Greeting\GetHandler as GreetingGetHandler;
+use RestDaemon\Example\Endpoint\Greeting\PostHandler;
+use RestDaemon\Example\Endpoint\Index\GetHandler;
+use RestDaemon\Example\Endpoint\Reusable\HelloHandler;
 
 return [
     // Endpoints in base (default or root) module: