+++ src/FreeElephants/RestDaemon/Endpoint/Handler/HandlerFactoryInterface.php
@@ -2,10 +2,12 @@
 
 namespace FreeElephants\RestDaemon\Endpoint\Handler;
 
+use Psr\Log\LoggerAwareInterface;
+
 /**
  * @author samizdam <samizdam@inbox.ru>
  */
-interface HandlerFactoryInterface
+interface HandlerFactoryInterface extends LoggerAwareInterface
 {
     public function buildHandler(string $className): EndpointMethodHandlerInterface;
 }