+++ src/FreeElephants/RestDaemon/Endpoint/Handler/InjectionHandlerFactory.php
@@ -1,13 +1,15 @@
 <?php
 
-
 namespace FreeElephants\RestDaemon\Endpoint\Handler;
 
-
 use FreeElephants\DI\Injector;
+use Psr\Log\LoggerAwareTrait;
+use Psr\Log\NullLogger;
 
 class InjectionHandlerFactory implements HandlerFactoryInterface
 {
+
+    use LoggerAwareTrait;
 
     /**
      * @var Injector
@@ -16,11 +18,14 @@ class InjectionHandlerFactory implements HandlerFactoryInterface
 
     public function __construct(Injector $injector)
     {
+        $this->logger = new NullLogger();
         $this->injector = $injector;
     }
 
     public function buildHandler(string $className): EndpointMethodHandlerInterface
     {
+        $this->logger->debug(sprintf('Build handler %s', $className));
+
         return $this->injector->createInstance($className);
     }
 }