+++ tests/cli/HelpCept.php
@@ -1,5 +1,5 @@
 <?php 
 $I = new CliTester($scenario);
-$I->runShellCommand('bin/rest-deamon');
+$I->runShellCommand('bin/rest-daemon');
 $I->seeInShellOutput('help');
 $I->seeResultCodeIs(0);