+++ CHANGELOG.md
@@ -9,7 +9,10 @@ Items with *BC!* note about backwards compatibility breaks!
 ## [Unreleased]
 
 ### Added
-- PSR-3 logging support and RequestLogger middleware. 
+- PSR-3 logging support and RequestLogger middleware.
+
+### Fixed
+- *BC!* deamon -> daemon mass typo in namespace and binary names. 
 
 ## [0.12.0] - 2018-02-11
 ### Added