+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Items with *BC!* note about backwards compatibility breaks!
 
 ## [Unreleased]
 
+### Removed
+- PHP 7.0 support
+
 ### Added
 - PSR-3 logging support and RequestLogger middleware.
 
@@ -16,6 +19,7 @@ Items with *BC!* note about backwards compatibility breaks!
 
 ### Fixed
 - *BC!* deamon -> daemon mass typo in namespace and binary names. 
+- Error on cli usage (added missing symfony/console dependency)
 
 ## [0.12.0] - 2018-02-11
 ### Added