+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Items with *BC!* note about backwards compatibility breaks!
 
 ## [Unreleased]
 
+## [0.13.0] 2019-05-26
 ### Removed
 - PHP 7.0 support
 
@@ -167,7 +168,8 @@ Items with *BC!* note about backwards compatibility breaks!
 - Adopt to use with one of two http-daemon drivers: Ratchet (ReactPHP) and Aerys (amphp). 
 - All features. 
 
-[Unreleased]: https://github.com/FreeElephants/rest-daemon/compare/0.12.0...HEAD
+[Unreleased]: https://github.com/FreeElephants/rest-daemon/compare/0.13.0...HEAD
+[0.13.0]: https://github.com/FreeElephants/rest-daemon/compare/0.12.0...0.13.0
 [0.12.0]: https://github.com/FreeElephants/rest-daemon/compare/0.11.0...0.12.0
 [0.11.0]: https://github.com/FreeElephants/rest-daemon/compare/0.10.0...0.11.0
 [0.10.0]: https://github.com/FreeElephants/rest-daemon/compare/0.9.0...0.10.0